Pull request rejects/fixes

PR rejects/fixes
